requirements.txt
opencv-python
mediapipe
pyautogui
pillow
pycaw
comtypes
